Frontend
Created dashboard layouts for Admin, Patient, and Practitioner.
Added simple table for showing list, for all practitioners and patients.
Added basic form with form fields to input data for new user(Still needs work for different users).

Both forms and tables needs to be worked on to connect to backend for CRUD.